Checkpoint

+++ helpers/parse_from_json.py
@@ -17,9 +17,9 @@ def parse(file_name):
     newCell = Cell(
       cell["top"],
       cell["left"],
-      int(splitH[0]),
+      float(splitH[0]),
       splitH[1],
-      int(splitV[0]),
+      float(splitV[0]),
       splitV[1],
       name
     )
